[EventDispatcher] Added TraceableEventDispatcher from HttpKernel
