+++ src/Symfony/Component/EventDispatcher/CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 2.5.0
 -----
 
+ * added Debug\TraceableEventDispatcher (originally in HttpKernel)
+ * changed Debug\TraceableEventDispatcherInterface to extend EventDispatcherInterface
  * added RegisterListenersPass (originally in HttpKernel)
 
 2.1.0
